Restructure into multiple modules
